pyproject.toml
[tool.poetry]
name = "keycloak_sync"
version = "0.1.0"
description = ""
authors = ["Kangyi LI <kangyi.li@polynom.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-keycloak = "^0.23.0"
configparser = "^5.0.1"
pandas = "^1.1.4"
numpy = "1.18.0"
google-cloud-storage = "^1.33.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.5.4"
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
